refactor(node): remove internal/errors.js
