+++ node/internal/readline/callbacks.js
@@ -21,10 +21,7 @@
 
 "use strict";
 
-import {
-  ERR_INVALID_ARG_VALUE,
-  ERR_INVALID_CURSOR_POS,
-} from "../../internal/errors.ts";
+import { ERR_INVALID_ARG_VALUE, ERR_INVALID_CURSOR_POS } from "../errors.ts";
 
 import { validateCallback } from "../validators.js";
