+++ node/internal/streams/state.js
@@ -1,7 +1,7 @@
 // Copyright Joyent and Node contributors. All rights reserved. MIT license.
 // deno-lint-ignore-file
 
-import { ERR_INVALID_ARG_VALUE } from "../../internal/errors.ts";
+import { ERR_INVALID_ARG_VALUE } from "../errors.ts";
 
 function highWaterMarkFrom(options, isDuplex, duplexKey) {
   return options.highWaterMark != null
