+++ node/internal/streams/transform.js
@@ -1,7 +1,7 @@
 // Copyright Joyent and Node contributors. All rights reserved. MIT license.
 // deno-lint-ignore-file
 
-import { ERR_METHOD_NOT_IMPLEMENTED } from "../../internal/errors.ts";
+import { ERR_METHOD_NOT_IMPLEMENTED } from "../errors.ts";
 import * as process from "../../_process/process.ts";
 import Duplex from "./duplex.js";
