+++ node/internal/streams/writable.js
@@ -17,7 +17,7 @@ import {
   ERR_STREAM_NULL_VALUES,
   ERR_STREAM_WRITE_AFTER_END,
   ERR_UNKNOWN_ENCODING,
-} from "../../internal/errors.ts";
+} from "../errors.ts";
 import * as process from "../../_process/process.ts";
 import destroyImpl from "./destroy.js";
 import EE from "../../events.ts";
